Safari-609.1.11.1/LayoutTests/js/function-dot-arguments-expected.txt

This test checks basic and corner cases of 'f.arguments'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS basicTest('one')[0] is 'one'
PASS basicTest('one')[2] is undefined.
PASS basicTest('one', 'two', 'three')[1] is 'two'
PASS lengthTest() is '0'
PASS lengthTest('From', '%E5%8C%97%E4%BA%AC', 360, '/', 'webkit.org') is '5'
PASS assignTest().toString() is '[object Arguments]'
PASS assignVarUndefinedTest().toString() is '[object Arguments]'
PASS assignVarUndefinedTest2().toString() is '[object Arguments]'
PASS assignVarInitTest().toString() is '[object Arguments]'
PASS assignVarInitTest2().toString() is '[object Arguments]'
PASS assignConstUndefinedTest().toString() is '[object Arguments]'
PASS assignConstUndefinedTest2().toString() is '[object Arguments]'
PASS assignConstInitTest().toString() is '[object Arguments]'
PASS assignConstInitTest2().toString() is '[object Arguments]'
PASS assignForInitTest().toString() is '[object Arguments]'
PASS assignForInitTest2().toString() is '[object Arguments]'
PASS assignForInInitTest().toString() is '[object Arguments]'
PASS paramInitTest(true).toString() is '[object Arguments]'
PASS paramFunctionConstructorInitTest(true).toString() is '[object Arguments]'
PASS tearOffTest()[0] is true
PASS tearOffTest2(true)[0] is true
PASS tearOffTest3(1, 2, 3, false) is [10, 2, 3, false]
PASS tearOffTest3a(1, 2, 3, false) is [10, 2, 3, false]
PASS tearOffTest3b(1, 2, 3, false) is [1, 2, 3, false]
PASS tearOffTest3c(1, 2, 3, false) is [1, 2, 3, false]
PASS tearOffTest4(1, 2, 3, false) is [10, 2, 3, false]
PASS tearOffTest4a(1, 2, 3, false) is [10, 2, 3, false]
PASS tearOffTest4b(1, 2, 3, false) is [1, 2, 3, false]
PASS tearOffTest4c(1, 2, 3, false) is [1, 2, 3, false]
PASS tearOffTest5(1, 2, 3, false) is [1, 2, 3, false]
PASS tearOffTest5a(1, 2, 3, false) is [1, 2, 3, false]
PASS tearOffTest5b(1, 2, 3, false) is [1, 2, 3, false]
PASS tearOffTest5c(1, 2, 3, false) is [1, 2, 3, false]
PASS tearOffTest6(1, 2, 3, false) is [1, 2, 3, false]
PASS tearOffTest6a(1, 2, 3, false) is [1, 2, 3, false]
PASS tearOffTest6b(1, 2, 3, false) is [1, 2, 3, false]
PASS tearOffTest6c(1, 2, 3, false) is [1, 2, 3, false]
PASS tearOffTest7(1, 2, 3, false) is [10, 2, 3, false]
PASS tearOffTest7a(1, 2, 3, false) is [10, 2, 3, false]
PASS tearOffTest7b(1, 2, 3, false) is [10, 2, 3, false]
PASS tearOffTest7c(1, 2, 3, false) is [10, 2, 3, false]
PASS tearOffTest8(1, 2, 3, false) is [10, 2, 3, false]
PASS tearOffTest8a(1, 2, 3, false) is [10, 2, 3, false]
PASS tearOffTest8b(1, 2, 3, false) is [10, 2, 3, false]
PASS tearOffTest8c(1, 2, 3, false) is [10, 2, 3, false]
PASS tearOffTest9(1, 2, 3, false) is [undefined, 2, 3, false]
PASS tearOffTest9a(1, 2, 3, false) is [undefined, 2, 3, false]
PASS tearOffTest9b(1, 2, 3, false) is [undefined, 2, 3, false]
PASS tearOffTest9c(1, 2, 3, false) is [undefined, 2, 3, false]
PASS tearOffTest10(1, 2, 3, false) is [undefined, 2, 3, false]
PASS tearOffTest10a(1, 2, 3, false) is [undefined, 2, 3, false]
PASS tearOffTest10b(1, 2, 3, false) is [undefined, 2, 3, false]
PASS tearOffTest10c(1, 2, 3, false) is [undefined, 2, 3, false]
PASS lexicalArgumentsLiveRead1(0, 2, 3) is 0
PASS lexicalArgumentsLiveRead2(1, 0, 3) is 0
PASS lexicalArgumentsLiveRead3(1, 2, 0) is 0
PASS lexicalArgumentsLiveWrite1(0, 2, 3) is 0
PASS lexicalArgumentsLiveWrite2(1, 0, 3) is 0
PASS lexicalArgumentsLiveWrite3(1, 2, 0) is 0
PASS argumentsNotLiveRead1(0, 2, 3) is 0
PASS argumentsNotLiveRead2(1, 0, 3) is 0
PASS argumentsNotLiveRead3(1, 2, 0) is 0
PASS argumentsNotLiveWrite1(0, 2, 3) is 0
PASS argumentsNotLiveWrite2(1, 0, 3) is 0
PASS argumentsNotLiveWrite3(1, 2, 0) is 0
PASS argumentsIdentity() is true
PASS overwroteArgumentsInDynamicScope1() is true
PASS overwroteArgumentsInDynamicScope2() is true
PASS overwroteArgumentsInDynamicScope3().toString() is '[object Arguments]'
PASS successfullyParsed is true

TEST COMPLETE
